refactor(playgrond): create ChatMessage component
