+++ packages/frontend/src/pages/Playground.svelte
@@ -1,15 +1,7 @@
 <script lang="ts">
 import { conversations } from '../stores/conversations';
 import { studioClient } from '/@/utils/client';
-import {
-  isAssistantChat,
-  type ChatMessage,
-  type UserChat,
-  isPendingChat,
-  isUserChat,
-  type AssistantChat,
-  isSystemPrompt,
-} from '@shared/src/models/IPlaygroundMessage';
+import { isAssistantChat, isPendingChat, isUserChat } from '@shared/src/models/IPlaygroundMessage';
 import NavPage from '../lib/NavPage.svelte';
 import { playgrounds } from '../stores/playgrounds-v2';
 import { catalog } from '../stores/catalog';
@@ -17,6 +9,8 @@ import Button from '../lib/button/Button.svelte';
 import { afterUpdate } from 'svelte';
 import ContentDetailsLayout from '../lib/ContentDetailsLayout.svelte';
 import RangeInput from '../lib/RangeInput.svelte';
+
+import { default as ChatMessageComponent } from '../lib/conversation/ChatMessage.svelte';
 
 export let playgroundId: string;
 let prompt: string;
@@ -44,28 +38,6 @@ $: {
   } else {
     sendEnabled = true;
   }
-}
-
-const roleNames = {
-  system: 'System prompt',
-  user: 'User',
-  assistant: 'Assistant',
-};
-
-function getMessageParagraphs(message: ChatMessage): string[] {
-  if (isAssistantChat(message)) {
-    if (!isPendingChat(message)) {
-      return message.content?.split('\n') ?? [];
-    } else {
-      return message.choices
-        .map(choice => choice.content)
-        .join('')
-        .split('\n');
-    }
-  } else if (isUserChat(message) || isSystemPrompt(message)) {
-    return message.content?.split('\n') ?? [];
-  }
-  return [];
 }
 
 function askPlayground() {
@@ -97,17 +69,6 @@ afterUpdate(() => {
 async function scrollToBottom(element: Element) {
   element.scroll?.({ top: element.scrollHeight, behavior: 'smooth' });
 }
-
-function elapsedTime(msg: AssistantChat): string {
-  if (isPendingChat(msg)) {
-    return ((Date.now() - msg.timestamp) / 1000).toFixed(1);
-  } else if (!!msg.completed) {
-    return ((msg.completed - msg.timestamp) / 1000).toFixed(1);
-  } else {
-    // should not happen
-    return '';
-  }
-}
 </script>
 
 {#if playground}
@@ -124,26 +85,7 @@ function elapsedTime(msg: AssistantChat): string {
                     <ul>
                       {#each conversation?.messages as message}
                         <li>
-                          <div class="text-lg" class:text-right="{isAssistantChat(message)}">
-                            {roleNames[message.role]}
-                          </div>
-                          <div
-                            class="p-4 rounded-md"
-                            class:bg-charcoal-400="{isUserChat(message)}"
-                            class:bg-charcoal-800="{isSystemPrompt(message)}"
-                            class:bg-charcoal-900="{isAssistantChat(message)}"
-                            class:ml-8="{isAssistantChat(message)}"
-                            class:mr-8="{isUserChat(message)}">
-                            {#each getMessageParagraphs(message) as paragraph}
-                              <p>{paragraph}</p>
-                            {/each}
-                          </div>
-                          {#if isAssistantChat(message)}
-                            <div class="text-sm text-gray-400 text-right" aria-label="elapsed">
-                              {elapsedTime(message)} s
-                            </div>
-                          {/if}
-                          <div></div>
+                          <ChatMessageComponent message="{message}" />
                         </li>
                       {/each}
                     </ul>
